Use FromSql for *const str instead of String
This is to avoid error below when multiple backends are enabled within
diesel. There are 2 options, switch the generig *const str to String or
change the from_sql to use String. The first was chosen since it also
avoids some extra allocation.

cannot infer type for type parameter `DB` declared on the trait `FromSql`